Add files via upload
Hello everyone! I want to share that my project focuses on urinary lithiasis. It was developed using the MVC architecture in Flutter and includes features for both biochemists and urologists. Each patient has a complete record containing all their medical information, including a unique ID, a microscopic image of their condition, and a QR code for secure identification and signature.